+++ src/lib/punctuation/dash.js
@@ -4,32 +4,60 @@ export function replaceThreeHyphensWithEmDash(string, locale) {
 
 
 
-export function replaceTwoHyphensWithEnDash(string, locale) {
-	return string.replace(/(--)/g, "–");
-}
-
-
-
-export function replaceSpacedHyphenWithEmDash(string, locale) {
+export function replaceSpacedHyphenWithDash(string, locale) {
 	let pattern = "[" + locale.spaces + "][" + locale.hyphen + "][" + locale.spaces + "]";
 	let re = new RegExp(pattern, "g");
-	let replacement = locale.space + locale.emDash + locale.space;
+	let replacement = "";
+
+	switch (locale.locale) {
+		case "en-us":
+		case "rue":
+		case "sk":
+		case "cs":
+			replacement = locale.space + locale.emDash + locale.space;
+			break;
+		case "de-de":
+			replacement = locale.space + locale.enDash + locale.space;
+			break;
+	}
+
 	return string = string.replace(re, replacement);
 }
 
 
 
-export function replaceSpacedEnDashWithEmDash(string, locale) {
-	let pattern = "[" + locale.spaces + "][" + locale.enDash + "][" + locale.spaces + "]";
+/*
+	Some languages require en dash, some require em dash.
+	This function will consolidate the dashes based on the locale
+
+	@param {string} string — input text for identification
+	@returns {string} — output with an appropriate dash between spaces
+*/
+export function consolidateSpacedDashes(string, locale) {
+
+	let pattern = "[" + locale.spaces + "][" + locale.enDash + "|" + locale.emDash + "][" + locale.spaces + "]";
 	let re = new RegExp(pattern, "g");
-	let replacement = locale.space + locale.emDash + locale.space;
+	let replacement = "";
+
+	switch (locale.locale) {
+		case "en-us":
+		case "rue":
+		case "sk":
+		case "cs":
+			replacement = locale.space + locale.emDash + locale.space;
+			break;
+		case "de-de":
+			replacement = locale.space + locale.enDash + locale.space;
+			break;
+	}
+
 	return string = string.replace(re, replacement);
 }
 
 
 
-export function fixSpacesAroundEmDash(string, locale) {
-	let pattern = "[" + locale.spaces + "]?[" + locale.emDash + "][" + locale.spaces + "]?";
+export function fixDashSpacesBetweenWords(string, locale) {
+	let pattern = "[" + locale.spaces + "]?[" + locale.emDash + "|" + locale.enDash + "][" + locale.spaces + "]?";
 	let re = new RegExp(pattern, "g");
 	let replacement = "";
 
@@ -41,6 +69,9 @@ export function fixSpacesAroundEmDash(string, locale) {
 		case "sk":
 		case "cs":
 			replacement = locale.hairSpace + locale.emDash + locale.hairSpace;
+			break;
+		case "de-de":
+			replacement = locale.hairSpace + locale.enDash + locale.hairSpace;
 			break;
 	}
 
@@ -92,10 +123,9 @@ export function fixDashBetweenOrdinalNumbers(string, locale) {
 */
 export function fixDash(string, locale) {
 	string = replaceThreeHyphensWithEmDash(string, locale)
-	string = replaceTwoHyphensWithEnDash(string, locale);
-	string = replaceSpacedHyphenWithEmDash(string, locale);
-	string = replaceSpacedEnDashWithEmDash(string, locale);
-	string = fixSpacesAroundEmDash(string, locale);
+	string = replaceSpacedHyphenWithDash(string, locale);
+	string = consolidateSpacedDashes(string, locale);
+	string = fixDashSpacesBetweenWords(string, locale);
 	string = fixDashBetweenCardinalNumbers(string, locale);
 	string = fixDashBetweenOrdinalNumbers(string, locale);
 	return string;